==> Rmarkdown/0_CDFA_CSV_KernFilter.Rmd <==
---
title: "Kern Filter applied to original CDFA data"
output:
  html_document: default
  pdf_document: default
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

**The purpose of this Rmarkdown file is to filter the CDFA data for Kern County, then write an output CSV of the filtered data**

**Notes:**

The data being read in with the read_csv() function is the original CDFA data. In the original data, years 2013-2017 were separated by tabs in the same Excel (.xcls) sheet. Prior to being read in to this Rmarkdown document, each tab was made into it's own individual .csv file in Excel. Years 2018 and 2019 were received in PDF format. We used Adobe tools to convert the PDFs to .csv files. 

**Output:**

- CDFA data for years 2013 to 2019 for Kern County
- ../R_output/CSV/CDFA/kern_filter/Kern_CDFA_APNs_'year'.csv

---

#### 1. Read in CSVs, filter dataframe for Kern County, and write export csv

```{r kern_filter_loop, message=FALSE, warning=FALSE}
library(tidyverse)

# Specify years to be included in this filtering process
years = 2013:2019

# Output folder for the following for loop
dir.create("../R_output/CSV/cdfa_kernFilter/", recursive = T)

# Read in CSV of APNs for each year 'i'
for(i in years){
  
  # Read in specific year's CDFA CSV file
  tmp = read_csv(paste0("../R_input/CSV/CDFA/original/CDFA_",i,".csv")) %>% 
    as.data.frame() %>% 
    filter(`Site County` == "Kern")
  
  # Assign a simple, year-specific name to the dataframe. These can be used for a quick visual check within R, of the files being written to CSV.
  assign(paste0("cdfa_",i),tmp) 
  
  write_csv(
    eval( # Evaluate what the name refers to -- In combination these commands are telling are to grab a specific dataframe
      as.name( # Tell R that paste is a name
        paste0("cdfa_",i))), # Paste elements together
       
    paste0("../R_output/CSV/cdfa_kernFilter/cdfa_kern_",i,".csv"))
  
  # Remove the 'tmp' dataframe to ensure there will be no accidental 
  # cross over between years i.e. processing 'tmp' for the wrong year
  rm(tmp)
  
  }

```
