==> Rmarkdown/1_CDFA_APN_edits.Rmd <==
---
title: "CDFA APN edit"
output:
  html_document: default
  pdf_document: default
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

**The purpose of this Rmarkdown document is to edit CDFA APNs so that they are in a format that can be matched to APNs in the Kern County Parcel shapefile attribute table.**

**Notes:**   

  * APNs are self-reported by farmers to the CDFA, and exist in a wide variety of formats in the CDFA database *so*:
  * Before reading CSVs into this Rmd document, the APN entry format was edited in Excel:
      * Each individual entry was separated into it's own column (i.e. if two entries were entered into the same cell, they were separated, maintaining the same row but ending up in individual columns, so that each cell contained no more than one entry)
      * If APN #s didn't have dashes, a dash was added between the 3 and 4 digit and between the 6 and 7 digit. 
      
  * This code uses a function stored in the "1_APN_edit.R" file to edited APN values. The function takes the original APN values and puts them into a '123-456-78' format to match the format of those in the Kern County Parcel shapefile attribute table
  * APN edit methods were evaluated on the number of matches made between CDFA APNs and Kern County Parcel data, and the accuracy of those matches was evaluated based on comparison of the Permittee and Company columns, respectively. 
  * In some instances, the Permittee and Company columns did not match. Further research showed that this was often the result of parent companies owning smaller companies
  
**Output:**

* CSV files of CDFA original and edited APNs in R_input and R_output files 
  - ../R_output/CSV/CDFA/final_APN/CDFA_APNs_[YEAR].csv
  - ../R_input/CSV/CDFA/final_APN/CDFA_APNs_[YEAR].csv

---

```{r load_packages_and_data, message=FALSE}
library(tidyverse)
library(reshape2)
library(rgdal)

# Bring in the function stored in the '1_APN_edit.R' script. If editing that function, 
# check the box that says 'source on save' to upload that latest version of the function 
# in this Rmarkdown document
source("../R/1_APN_edit.R") 
```

#### Apply function to years 2013 to 2019. It was ultimately a little easier to manually work through each year than to apply a for-loop because the column indexes (apn_cols) were different for each year. 

#### 2013
```{r yr2013, message=FALSE}

# Read in csv that contains APNs for Kern County in 2013. Prior to being read in to this 
# Rmarkdown document, APNs may have been edited in excel in one or more of the following 
# ways: (1) if multiple  APNs in one cell, seperated into individual cells across multiple 
# columns, (2) dashes added to those numbers that didn't have them. If dashes were added I 
# always started counting from the left and added a dash after the 3rd number and after the 
# 6th number. Original versions of the APNs were kept for comparison
cdfa_2013_raw = read_csv("../R_input/CSV/CDFA/excel_edited_APNs/CDFA_excelAPNs_2013.csv")

# Define parameters for APN_edit function in a list of elements. These also could be defined 
# directly in the function arguments below. 
apn_params = list(id_cols = 1:2, # ID Column used to go from wide to long format
                  apn_cols = 19:72, # These are the APN columns what will be melted into one
                  col_names = c("orgID", "company","APN_index","original_apn"), # Column names of the new long table
                  year = 2013) # Year being evaluated

# Create output directories
# dir.create("../R_input/CSV/CDFA/final_APN/",recursive = T)
# dir.create("../R_output/CSV/CDFA/final_APN/",recursive = T)

# Use the APN_edit_fun() with the list of parameters for the 2013 APNs. 
cdfa_2013_final_df = APN_edit_fun(cdfa_2013_raw, # 2013 CDFA Dataframe
                                  id_cols = apn_params$id_cols, 
                                  apn_cols = apn_params$apn_cols, 
                                  col_names = apn_params$col_names, 
                                  year = apn_params$year)

rm(apn_params)
# The warning message refers to empty cells that results when APNs are separated by '-' into 
# multiple columns. Those that have fewer than the maximum number of ? receive NAs in 
# those columns. The warning is not problematic and does not effect those APN values. 
```


All subsequent years follow the same order of operations as those explained in the 2013 code chunk above. 

#### 2014
```{r yr2014, message=FALSE}
cdfa_2014_raw = read_csv("../R_input/CSV/CDFA/excel_edited_APNs/CDFA_excelAPNs_2014.csv")

apn_params = list(id_cols = 1:2,
                  apn_cols = 21:74, 
                  col_names = c("orgID", "company","APN_index","original_apn"),
                  year = 2014)

cdfa_2014_final_df = APN_edit_fun(cdfa_2014_raw, 
                                  id_cols = apn_params$id_cols, 
                                  apn_cols = apn_params$apn_cols, 
                                  col_names = apn_params$col_names, 
                                  year = apn_params$year)
rm(apn_params)
```


#### 2015
```{r yr2015}
cdfa_2015_raw = read_csv("../R_input/CSV/CDFA/excel_edited_APNs/CDFA_excelAPNs_2015.csv")

apn_params = list(id_cols = 1:2, 
                  apn_cols = 21:74, 
                  col_names = c("orgID", "company","APN_index","original_apn"),
                  year = 2015)

cdfa_2015_final_df = APN_edit_fun(cdfa_2015_raw, 
                               id_cols = apn_params$id_cols, 
                               apn_cols = apn_params$apn_cols, 
                               col_names = apn_params$col_names, 
                               year = apn_params$year)
rm(apn_params)
```


#### 2016
```{r yr2016, message=FALSE}
cdfa_2016_raw = read_csv("../R_input/CSV/CDFA/excel_edited_APNs/CDFA_excelAPNs_2016.csv")

apn_params = list(id_cols = 1:2, 
                  apn_cols = 21:74, 
                  col_names = c("orgID", "company","APN_index","original_apn"),
                  year = 2016)

cdfa_2016_final_df = APN_edit_fun(cdfa_2016_raw, 
                                  id_cols = apn_params$id_cols, 
                                  apn_cols = apn_params$apn_cols, 
                                  col_names = apn_params$col_names, 
                                  year = apn_params$year)
rm(apn_params)
```


#### 2017
```{r yr2017, message=FALSE}
cdfa_2017_raw = read_csv("../R_input/CSV/CDFA/excel_edited_APNs/CDFA_excelAPNs_2017.csv")

apn_params = list(id_cols = 1:2, 
                  apn_cols = 21:74, 
                  col_names = c("orgID", "company","APN_index","original_apn"),
                  year = 2017)

cdfa_2017_final_df = APN_edit_fun(cdfa_2017_raw, 
                                  id_cols = apn_params$id_cols, 
                                  apn_cols = apn_params$apn_cols, 
                                  col_names = apn_params$col_names, 
                                  year = apn_params$year)
rm(apn_params)
```


#### 2018
```{r yr2018, message=FALSE}
cdfa_2018_raw = read_csv("../R_input/CSV/CDFA/excel_edited_APNs/CDFA_excelAPNs_2018.csv")

apn_params = list(id_cols = 1:2, 
                  apn_cols = 13:35, 
                  col_names = c("orgID", "company","APN_index","original_apn"),
                  year = 2018)

cdfa_2018_final_df = APN_edit_fun(cdfa_2018_raw, 
                                  id_cols = apn_params$id_cols, 
                                  apn_cols = apn_params$apn_cols, 
                                  col_names = apn_params$col_names, 
                                  year = apn_params$year)

rm(apn_params)
```

#### 2019
```{r yr2019, message=FALSE}
cdfa_2019_raw = read_csv("../R_input/CSV/CDFA/excel_edited_APNs/CDFA_excelAPNs_2019.csv")

apn_params = list(id_cols = 1:2, 
                  apn_cols = 13:35, 
                  col_names = c("orgID", "company","APN_index","original_apn"),
                  year = 2019)

cdfa_2019_final_df = APN_edit_fun(cdfa_2019_raw, 
                                  id_cols = apn_params$id_cols, 
                                  apn_cols = apn_params$apn_cols, 
                                  col_names = apn_params$col_names, 
                                  year = apn_params$year)
rm(apn_params)
```


#### Counting total APNs 
```{r counting_APNS}
# Years to count from
years = 2013:2019

# Initiate data.frame to store results in for loop
all_apns = as.data.frame(matrix(ncol = 2))

# Name data.frame columns
colnames(all_apns) = c("unique_apn","year")

# Being for loop that combines unique APN entries for all years. The length of the final table == the total number of unique APNs. A unique APN in this context is actually an APN-year combination, as the same APN value represents different data for different years. 
for(i in years){
  
  print(i)
  tmp = eval(as.name(paste0("cdfa_",i,"_final_df")))
  tmp = data.frame(unique_apn = unique(tmp$edited_APN))
  tmp$year = i
  all_apns = rbind(all_apns,tmp)
  rm(tmp)
  
}
```
