==> Rmarkdown/2_KernPrcl_CDFA_Join_and_CDFA_sctn_ID.Rmd <==
---
title: "Kern County Parcel Shapefile -- CDFA table join on APN Field and CDFA TRS section ID"
output:
  pdf_document: default
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

**The purpose of this Rmarkdown document is to join CDFA tabular data to the Kern County Parcel shapefile. Two previous Rmarkdown documents prepared APNs for the join**

**Notes:** 

* This Rmarkdown document uses a function in the "2_KernCountyParcel_CDFA_join.R" file to join CDFA and Kern County Parcel data on their respective APN columns. See notes in that R script for more details about that function. It also identifies the PLSS sections in Kern County that have TRS codes (from COMTRS code) that match those in CDFA data which were reported in place of APNs. 

**Output:**

* Shapefile of with CDFA data joined to Kern County Parcel shapefile attribute table by APN
    - ../R_output/spatial/CDFA_APN_parcels/[YEAR]/cdfa_parcels_[YEAR].shp
    
* Shapefile of PLSS sections identified using TRS codes reported in place of APNs in CDFA data
    - ../R_output/spatial/CDFA_PLSS_sections/cdfa_sections_[YEAR].shp
---

#### 1. Load packages, function, and data
```{r load_packages, message=FALSE}
library(tidyverse) 
library(rgdal) 
library(sf) 

# Load function stored in the '3_KernCountyParcel_CDFA_join.R' script. 
source("../R/2_KernCountyParcel_CDFA_join.R")

# Read in Kern County Parcel data and transform Kern County Parcel data into the correct CRS for project. CA Teale Albers 2011
KernCoParcels <- readOGR("../R_input/spatial/Kern_Parcels_2017_Final/Kern_Parcels_2017_Final.shp") %>% 
  spTransform(., CRS("+proj=aea +lat_1=34 +lat_2=40.5 +lat_0=0 +lon_0=-120 +x_0=0 +y_0=-4000000 +ellps=GRS80 +datum=NAD83 +units=m +no_defs "))

```

#### 2. Apply function that joins CDFA tabular data to Kern County Parcel spatial data on the APN field for specified years: 2013 - 2019. 
```{r apply_function}

# Specify years 
yrs = 2013:2019

# Create output directory
dir.create("../R_output/spatial/CDFA_APN_parcels/", recursive = T)

# For more notes about this function see the .R file
for(i in yrs){
  
  KernCoParcel_CDFA_join(year = i, # Year index
                         kern_parcel_shp = KernCoParcels) # Shapefile to be used

  }
```

#### 3. Identify PLSS Sections that contain organic parcels (the next phase of this, finding WHICH parcels are organic, is in RMD 4)

##### A) Bring in Kern County PLSS Sections shapefile
```{r PLSS_shapefile}

# Bring in Kern Sections shapefile
kern_sections = readOGR("../R_input/spatial/Kern_sections/Kern_15r.shp")%>% 
  spTransform(., CRS("+proj=aea +lat_1=34 +lat_2=40.5 +lat_0=0 +lon_0=-120 +x_0=0 +y_0=-4000000 +ellps=GRS80 +datum=NAD83 +units=m +no_defs ")) %>% 
  st_as_sf()

# Filter attribute table for necessary fields/columns
kern_sections_slct = kern_sections %>% 
  dplyr::select(AREA, MTRS)

# Create TRS column by trimming "m' off the beginning of the MTRS column values
kern_sections_slct$trs <- str_sub(kern_sections_slct$MTRS,2,9)

```

##### B) Read in CDFA PLSS csv, join to PLSS shapefile and write output shapefile
```{r CDFA_trs}

# Years to evaluate
yrs = 2013:2019

# Initiate empty dataframe where results will be stored
all_trs = data.frame(matrix(ncol = 2))
# Name output df columns
colnames(all_trs) = c("uniq_trs","year")

for(i in yrs){
  
  # Assign output file name variable for specific year
  outputfile = paste0("../R_output/spatial/CDFA_PLSS_sections/",i)
  
  # Create output folder
  dir.create(outputfile)
  
  # Read in CDFA data with the TRS codes pulled from the APN and address columns
  cdfa_trs_df = read.csv(paste0("../R_input/CSV/CDFA/excel_edited_PLSS/cdfa_trs_",i,".csv")) %>% 
    dplyr::select("Company","t","r","s") 
  
  # Remove rows that contain NA values in the T, R, or S columns
  cdfa_trs_df = filter(cdfa_trs_df,(!is.na(t)|!is.na(r)|!is.na(s)))
  
  # If a string in the 's' column is less than 2 digits long, pad on the left side with a 0
  cdfa_trs_df$s <- str_pad(cdfa_trs_df$s, width=2, side="left", pad="0")
   
  # If a string in the 's' column is less than 2 digits long, pad on the left side with a 0
  cdfa_trs_df$t <- str_pad(cdfa_trs_df$t, width=3, side="left", pad="0")
  
  # If a string in the 's' column is less than 2 digits long, pad on the left side with a 0
  cdfa_trs_df$r <- str_pad(cdfa_trs_df$r, width=3, side="left", pad="0")
  
  # Create the complete 'trs' column
  cdfa_trs_df$trs = paste0(cdfa_trs_df$t,cdfa_trs_df$r,cdfa_trs_df$s)
  
  # Get a count of how many times each TRS is repeated and how many unique codes there are
  cdfa_trs_summary = cdfa_trs_df %>%
    group_by(Company,trs) %>%
    summarise(count = n())
  
  # Uncomment the following line to print number of unique TRS codes by year
  # print(i,table(cdfa_trs_summary$Company))
  
  # Merge CDFA to PLSS sections on the TRS columns created above
  cdfa_sections = merge(kern_sections_slct,cdfa_trs_df,by = "trs")
  
  # The merge produces some duplicate matches. This gets rid of them, then cleans up the data.frame and turns it back into a Spatial Polygons file
  cdfa_sections_unique = distinct(cdfa_sections,.keep_all = T) %>% 
    dplyr::select(trs,company = "Company",geometry) %>% 
    as(.,"Spatial")
  
  # Define the layer name for year i
  cdfalayer = paste0("cdfa_sections_",i)
  
  # Write output file for year i
  # writeOGR(cdfa_sections_unique,
  #          dsn = outputfile,
  #          layer = cdfalayer,
  #          driver = "ESRI Shapefile",
  #          overwrite = TRUE)
  
  # Remove CDFA data frames to avoid crossover between years
  rm(cdfa_sections)
  rm(cdfa_sections_unique)
  rm(cdfa_trs_summary)
  
  # Data.frame of unique TRS codes for year i
  cdfa_trs_df = data.frame(uniq_trs = unique(cdfa_trs_df$trs))
  cdfa_trs_df$year = i
  
  #Compile dataframe of unique TRS codes for each year. 
  all_trs = rbind(all_trs,cdfa_trs_df)
  rm(cdfa_trs_df)
  
}

```
