==> Rmarkdown/4_KernAg_CDFA_sptlJoin.Rmd <==
---
title: "Kern County Agriculture -- CDFA parcel spatial join"
output:
  html_document: default
  pdf_document: default
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

**The primary purpose of this Rmarkdown file is to identify organic agriculture fields using 3 approaches:**

1. Spatial join between CDFA APN-identified parcels (from RMD 2) and Kern Ag parcels
2. Spatial join between CDFA PLSS sections (identified using TRS codes in CDFA csv in RMD 2) and Kern Ag parcels
3. Filtering the commodity [COMM] column in the Kern Agriculture Shapefile for the strings "ORG" or "ORGANIC" 

**Notes:**

* CDFA parcels were determined by formatting APNs reported to the CDFA by growers, then matching the formatted APNs to Kern County Assessor Parcel spatial data on the APN field. 
* Kern County Agriculture parcels are available on the Kern County Website. 
* Polygons in the Kern County Agriculture shapefile and Kern County Parcel shapefile have different dimensions, with Kern Agriculture fields generally having smaller geometries. This means that one Kern County parcel often overlaps and spatially joins with multiple Kern County Ag fields, which results in false positives (conventional fields IDed as organic based on this spatial join). Subsequent data processing dealt with false positives by evaluating pesticide use on each field, and designating fields as conventional if they had used any inorganic pesticide products. 

* This Rmarkdown document uses a function in the "4_KernAg_CDFA_Join.R" file. Function variables include:
       * Year (year)
       * Buffer width (buf_width)
       * Two true/false argument for function output

**Output:**
       - ../R_output/spatial/final_organics_data/[YEAR]/all_organics_[YEAR].shp
       
---

#### 1. Load packages, set working diretory, and source function in the KernAg_CDFA_join.R file 
```{r packages_and_function}
library(tidyverse) 
library(rgdal) 
library(rgeos) 
library(sf) 
library(raster)
library(cleangeo)

# Read in function in .R document that performs a spatial join between Kern County Agriculture and CDFA parcel shapefiles. 
# See the .R file for more information about the function
source("../R/4_KernAg_CDFA_Join.R")
```

#### 2. Implement function that performs a spatial join between APN-identified CDFA parcels and Kern Agriculture parcels for specific years and desired buffer width. See notes in the following code chunk for more information about buffer widths and how 50m was determined
```{r implement_function, message=FALSE, warning=FALSE}
# Years of interest
years = 2013:2019

# Buffer widths: In the function these are negative and reduce the CDFA field size. 
# A larger buffer results in a more conservative spatial join between the CDFA parcels and 
# Kern County Agriculture Parcels. The fields in each shapefile have different dimensions, and 
# with no buffer applied CDFA fields match to Kern Ag parcels that hardly touch, increasing the 
# risk of false matches. 

# With a 100-meter buffer, the Kern Ag fields that CDFA parcels join to are almost entirely 
# within the CDFA parcels which results in a loss of reasonable matches.

# A 50-meter buffer provides conservative but seemingly accurate matches. 
# Matches were checked by comparing the company and permittee columns in the attribute 
# table of the resulting joined shapefile.

buf_width = c(50)

for(i in years){
  
  for(j in buf_width){
  
    # Create output directory for the output CSV file that lists the CDFA data 
    # with commodities from Kern Ag data attached
    dir.create(paste0("../R_output/CSV/cdfa_kernAg_join/",i),
               recursive = TRUE)
    
    # Create the output directory for the full CDFA_KernAg_join shapefile, no filters applied. 
    dir.create(paste0("../R_output/spatial/CDFA_KernAg_join/",i,"/buffer",j),
               recursive = TRUE)
    
    # The function output can be:
    #    A joined shapefile that includes all AG parcels and the Kern Ag/CDFA attribute table
    #    (write_all_shp = TRUE)
    #    A joined shapefile that is filtered to only include ag parcels that joined to a CDFA parcel
    #    (write_cdfa_shp = TRUE)
    
    # Apply the function that is in the KernAg_CDFA_Join.R script
    kernAg_CDFA_joinFun(i, buf_width = j, write_all_shp = T, write_cdfa_shp = T)
 
     }
  }

```

#### 3. Perform spatial join between CDFA PLSS sections and Kern County Agriculture Parcels -- Extract Ag parcels that overlap with CDFA sections. 
```{r trs_ag_join, warning=FALSE}
# Years to evaluate
years = 2013:2019

# Initiate dataframe we will use to tally the number of organic observations using each CDFA org designation
cdfa_orgs_table = matrix(ncol = 4, nrow = length(years))

# Name columns
colnames(cdfa_orgs_table) = c("year","apn","trs","total")

# Fill in years column
cdfa_orgs_table[,1] <- c(years)

# Create output folder if not already present
dir.create("../R_output/spatial/final_organics_data/")

for(i in years){
  
  # Read in Kern County Ag-CDFA shapefile produced on function above for year i 
   tmp <- readOGR(paste0("../R_output/spatial/CDFA_KernAg_join/",i,"/buffer50/KernAg_CDFA_all_",i,"_B50.shp")) %>% 
     cleangeo::clgeo_Clean(.) %>% 
     st_as_sf()
  
   # Assign unique name and remove tmp file
  assign(paste0("cdfa_kern_apn_",i),tmp)
  rm(tmp)
  
  ## Read in Kern County Ag data with soil quality included attribute table, produced in RMD 3
  tmp <- read_sf(paste0("../R_output/spatial/KernAg_withSoil/",i,"/KernAg_withSoil_",i,".shp")) %>%
    dplyr::select(PERMIT,PMT_SITE,S_STATUS,P_STATUS,COMM,PERMITTEE,AG_TRS) %>% # Select columns 
    as(.,"Spatial") %>% # Convert to shapefile
    cleangeo::clgeo_Clean(.) %>% # Clean geometries
    st_as_sf() # Convert back to spatial data.frame
  
  # Assign unique name and remove tmp file
  assign(paste0("kern_",i),tmp)
  rm(tmp)
  
  ##  Read in shapefile of the PLSS Sections that joined to CDFA data on TRS codes. 
  tmp <- read_sf(paste0("../R_output/spatial/CDFA_PLSS_sections/",i,"/cdfa_sections_",i,".shp")) %>% 
    st_as_sf() %>% 
    mutate(trs_match = 1) %>% 
    as(.,"Spatial") %>% 
    cleangeo::clgeo_Clean(.) %>% 
    st_as_sf()
  
  # Assign unique name and remove tmp file
  assign(paste0("cdfa_sections_",i),tmp)
  rm(tmp)
  
  # Join Kern County Agricultural data to CDFA sections
  cdfasection_ag_join <- st_join(eval(as.name(paste0("kern_",i))), # Kern ag data for year i
                      eval(as.name(paste0("cdfa_sections_",i))), # CDFA PLSS sections for year i
                      largest=T) # Join Polygons with the largest overlap
  
  # Create binary column where 1 is assigned to rows with the word "ORGANIC" in the permittee name, and otherwise 0
  cdfasection_ag_join$trs_orgPermittee_match <- ifelse((str_detect(cdfasection_ag_join$PERMITTEE,"ORGANIC") &
                                                          cdfasection_ag_join$trs_match == 1),1,0)
    
  # Select columns wanted in attribute table
  org_cdfaSection_ag  = cdfasection_ag_join %>% 
  dplyr::select("permittee" = PERMITTEE,
               company,
               trs_match,
               trs_orgPermittee_match)
  
  # Join TRS section data to Kern Ag/CDFA_APN data. 
  apn_trs_organics_join <- st_join(eval(as.name(paste0("cdfa_kern_apn_",i))), # Kern Ag-CDFA.APN data for year i
                      org_cdfaSection_ag, # Kern Ag-CDFA.PLSS data for year i 
                      largest=T,
                      suffix = c(".apn",".trs"))
  
  # Select columns wanted in final attribute table and rename where necessary
  apn_trs_organics_final = apn_trs_organics_join %>% 
    dplyr::select("permit" = PERMIT,
         "permitsite" = PMT_SIT,
         "s_status" = S_STATU,
         "p_status" = P_STATU,
         "permittee" = PERMITT,
         "comm" = COMM,
         "comm_code" = COMM_CO,
         "agroclass" = AGROCLA,
         "genus" = GENUS,
         "family" = FAMILY,
         "comm_edit"=COMM_nw,
         "acres"=ACRES,
         "soil_wNA"=SOIL_wNA,
         "soil_noNA" = SOIL_nNA,
         "soil_chem" = SOIL_ch,
         "soil_temp" = SOIL_tm,
         "soil_hydr" = SOIL_hy,
         "apn"=CDFA,
         "trs_only" = trs_match,
         "trs_orgPermittee" = trs_orgPermittee_match)
  
    # Replace NAs with 0s in the trs_only and trs_orgPermittee column to make binary 
    apn_trs_organics_final$trs_only[is.na(apn_trs_organics_final$trs_only)] = 0
    apn_trs_organics_final$trs_orgPermittee[is.na(apn_trs_organics_final$trs_orgPermittee)] = 0
  
    # Create a CDFA binary organic column based on the apn_match and trs_match columns. If either of those indicates that the parcel is organic then the value in the CDFA_org column == 1.
    apn_trs_organics_final$cdfa_organic <- ifelse(apn_trs_organics_final$apn==1|apn_trs_organics_final$trs_only==1, # If APN or TRS organic
                                              1, # put a 1 in the cdfa_organic column
                                              0) # otherwise put a 0
    
    # Assign unique name to dataframe
    assign(paste0("cdfa_orgs_",i),apn_trs_organics_final)
    
      # # Uncomment to write intermediate Output File
      # writeOGR(obj = as(apn_trs_organics_final,"Spatial"),
      #    dsn = paste0("../R_output/spatial/CDFA_KernAg_join/",i,"/buffer50"),
      #    layer = paste0("TRSandAPN_organics_",i),
      #    driver = "ESRI Shapefile",
      #    overwrite_layer = T)
 
     # Write table compiled across years that shows the number of fields IDed as organic for each method 
     tmp_sum <- data.frame(year = i,
                        apn = sum(apn_trs_organics_final$apn),
                        trs_only = sum(apn_trs_organics_final$trs),
                        total = sum(apn_trs_organics_final$cdfa_organic))
  
     # Merge about with the compiled years df
      cdfa_orgs_table = merge(cdfa_orgs_table,tmp_sum) # Merge previous years with year i
      rm(tmp_sum)                      
  }

```

#### 4. Identify organic fields based on Kern Ag "COMM" column
```{r COMM_organics}

years = 2013:2019

for(i in years){
  
  # Get file with APN and TRS organic designations
  tmp = eval(as.name(paste0("cdfa_orgs_",i)))
  
  # Create an 'organic commodity' field. If contains "ORG" but not "SORGHUM" == 1. This would exclude organic sorghum, but haven't seen any of that. 
  tmp$comm_org <- ifelse((str_detect(tmp$comm,"ORG")&!str_detect(tmp$comm,"SORG")),1,0)
  
  # Create Ogranic permittee field
  tmp$permittee_org <- ifelse(str_detect(tmp$permittee,"ORGA"),1,0)
  
  # Create a combined organic field that incorporates comm_org, trs_org, and apn_org
  tmp$all_organic = ifelse((tmp$comm_org==1|tmp$trs_only==1|tmp$apn==1),1,0)
  
  print(c(i,
          "comm_org", sum(tmp$comm_org, na.rm = T),
          "permittee_org",sum(tmp$permittee_org),
          "apn_org",sum(tmp$apn),
          "trs_org",sum(tmp$trs_only),
          "all_org", sum(tmp$all_organic)))
  
  # Uncomment to write intermediate output file
  # writeOGR(obj = as(tmp,"Spatial"),
  #          dsn = paste0("../R_output/spatial/CDFA_KernAg_join/",i,"/buffer50"),
  #          layer = paste0("all_organics_",i),
  #          driver = "ESRI Shapefile",
  #          overwrite_layer = T)
  
  # Write output file with data compiled from: CDFA tabular data, Kern County Agriculture data, soil data
  writeOGR(obj = as(tmp,"Spatial"),
           dsn = paste0("../R_output/spatial/final_organics_data/",i),
           layer = paste0("all_organics_",i),
           driver = "ESRI Shapefile",
           overwrite_layer = T)
  
  # Assign unique name for year i and remove tmp file
   assign(paste0("all_orgs_",i),tmp)
   rm(tmp)

}


```
