==> Rmarkdown/6_CompileYears_and_Explore.Rmd <==
---
title: "6. Compile data and explore"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

**The purpose of this Rmarkdown file is to compile the finalized datasets for all years, and write output tabular and shapefiles**  

**Output:**
- ../R_output/CSV/compiled_organics1319.dta
- ../R_output/CSV/compiled_organics1319.csv
- ../R_output/spatial/final_organics_data/all_years/all_years_organics.shp

---

#### 1. Get Packages, set working directory, etc. 
```{r packages}
library(tidyverse)
library(rgdal)
library(sf)
library(haven)
# library(ggplot2) Isn't this part of tidyverse?

# Read in pesticide use data
pest_data_raw = read_dta("../R_input/dta/KernFieldCrops1319.dta") %>% 
  as.data.frame()
```

#### 2. Read in finalized data (shapefile) for each year, and turn into dataframe to rbind in next chunk
```{r data}

years = 2013:2019

## Read in shapefiles and make into a simple dataframe without geometry
for(i in years){
  
  tmp = read_sf(paste0("../R_output/spatial/final_organics_data/",i,"/all_organics_pest_",i,".shp")) 
  
  assign(paste0("organics_df_",i),tmp)
  
  rm(tmp)
  
}
colnames(organics_df_2013)
```

#### 3. Redo agroclass join to 2019
```{r fix_2019_agroclassJoin}

##### Fix the rows with '-' at the end of the comm column #####

# Convert comm_edit column from factor to character
organics_df_2019$comm_dt=as.character(organics_df_2019$comm_dt)

# Separate edited comm column (comm_dt) by dash
organics_df_2019 = organics_df_2019 %>%
    separate(col = "comm_dt",
             into = c("COMM_x","COMM_y"),
             sep = "-",
             remove = FALSE)

# Get rid of '-' where it exists at the end of commodity strings
organics_df_2019$comm_dt = ifelse((organics_df_2019$COMM_y==""| # If there is a 0 , "", or NA in the comm_y column from above
                                     organics_df_2019$COMM_y=="O"|
                                     is.na(organics_df_2019$COMM_y)),
                                  organics_df_2019$COMM_x, # Then comm equals the value in comm_x
                                  organics_df_2019$comm_dt) # Otherwise it equals the comm_dt column as is

## Read in agroclass data
agro_class <- read_csv("../R_input/CSV/Kern_AG/comm_subgroups.csv")

## Find 2019 rows that have NA in the agroclass column
missing_agrcls = filter(organics_df_2019,is.na(agrclss))
# 218 before

## Rejoin 2019 dataframe with updated comm column to agroclass data.set
organics_df_2019 = left_join(x = organics_df_2019,y = agro_class, by = c("comm_dt" = "commodity"))

# See how many still have NA values
missing_agrcls = filter(organics_df_2019,is.na(agroclass))
# 141 after

# Look at what crop types have NA values
data.frame(table(organics_df_2019$comm_dt[is.na(organics_df_2019$agroclass)])) %>% 
  arrange(-Freq)

# Select columns and rename where necessary
organics_df_2019 = organics_df_2019 %>% 
  dplyr::select(permtst,
                permit,
                s_stats,
                p_stats,
                permitt,
                commdty,
                comm_cd,
                "agrclss" = agroclass, 
                "genus" = genus.y, 
                "family" = family.y, 
                comm_dt,
                acres,
                sol_wNA,sol_nNA,sol_chm,sol_tmp,sl_hydr,
                apn,trs_nly,
                trs_rgP,
                cdf_rgn,
                comm_rg,
                prmtt_r,
                all_rgn,
                year,
                permitn,
                locatin,
                N_crps_,
                hectars, 
                geometry)

# Check that columns are in the same order and have the same names are other years, for the rbind
colnames(organics_df_2013) == colnames(organics_df_2019)
```

#### 4. CSV OUTPUT: Compile (rbind) years into one df
```{r compile_for_csv}

# Initiate dataframe using 2013 dataset
all_years_df <- data.frame(matrix(ncol=ncol(organics_df_2013)-1,nrow=0))

# Rename columns to match 2013 dataset
colnames(all_years_df) <- colnames(organics_df_2013)[1:ncol(organics_df_2013)-1]

# Rbind all years
for(i in years){
  
  # Dataframe for year i
  tmp <- eval(as.name(paste0("organics_df_",i))) %>% 
    as.data.frame() %>% 
    dplyr::select(-geometry)
  
  # Fill in year column 
  tmp$year = i
  
  # Rbind to final dataframe
  all_years_df = rbind(all_years_df,tmp)
  
  # Remove tmp data.frame
  remove(tmp)
  
}

# Look at number of rows for each year
table(all_years_df$year)
# 2013  2014  2015  2016  2017  2018  2019 
# 13997 13906 14057 14017 14287 14431 14841 

## Rename columns with full names
colnames(all_years_df) = c("permitsite","ag_permit","s_status","p_status","permittee","comm_original","comm_code","agroclass","genus","family",
                           "comm_edit","acres","soil_NA","soil_noNA","soilX_chem","soilX_temp","soilX_hydro","apn_org","trsonly_org",
                           "trspermittee_org","cdfa_org","kerncomm_org","kernpermittee_org","all_org","year","pur_permit","pur_site","n_crops","hectares")

# Write output dta and csv files
write_dta(all_years_df, "../R_output/CSV/compiled_organics1319.dta")
write_csv(all_years_df, "../R_output/CSV/compiled_organics1319.csv")

```

#### 5. Do the same as above without removing geometry then compile into single shapefile
```{r compile_for_shp}

# Initiate dataframe using 2013 dataset
all_years_shp <- organics_df_2013
all_years_shp$year = 2013 # Fill in year

# Years to loop through
years = 2014:2019 

for(i in years){
  
  # Get dataframe for year i
  tmp <- eval(as.name(paste0("organics_df_",i)))
  
  # Fill in year
  tmp$year = i
  
  # R bind to compiled dataset
  all_years_shp = rbind(all_years_shp,tmp)
  
  # Remove tmp dataframe
  remove(tmp)
  
}

# Check the number of rows in each year
table(all_years_shp$year)

# Rename columns
colnames(all_years_shp) = c("permitsite","ag_permit","s_status","p_status","permittee","comm_original","comm_code","agroclass","genus","family",
                           "comm_edit","acres","soil_NA","soil_noNA","soilX_chem","soilX_temp","soilX_hydro","apn_org","trsonly_org","trspermittee_org",
                           "cdfa_org","kerncomm_org","kernpermittee_org","all_org","year","pur_permit","pur_site","n_crops","hectares","geometry")

# Write output shapefile
writeOGR(as(all_years_shp,"Spatial"),
         dsn = "../R_output/spatial/final_organics_data/all_years",
         layer = "all_years_organics",
         driver = "ESRI Shapefile",
         overwrite_layer = T)
```


#### 5. Visualize some data 

Organics per year
```{r orgs_per_year}

org_fields_perYear = all_years_df %>% 
  group_by(year) %>% 
  summarise(cdfa_organics = sum(cdfa_org),
            all_organics = sum(all_org))

ggplot(org_fields_perYear, aes(x = as.factor(year), y = cdfa_organics, fill = as.factor(year)))+
  geom_col()+
  geom_label(aes(label = cdfa_organics))+
  theme_classic()

ggplot(org_fields_perYear,aes(x = as.factor(year),y = all_organics,fill = as.factor(year)))+
  geom_col()+
  geom_label(aes(label = all_organics))+
  theme_classic()
```

Field size (hectares) -- Get into same figure
```{r field_size}
mean_hectares = all_years_df %>% 
  group_by(year,cdfa_org) %>% 
  summarise(mean_hectares = round(mean(hectares),2))

ggplot(mean_hectares,aes(x = year,y = mean_hectares,fill=as.factor(cdfa_org)))+
  geom_col(position = "dodge")+
  # geom_label(aes(label = mean_hectares)) + 
  theme_classic()+
  scale_fill_manual(values = c("darkgoldenrod2", "#2e4057"), labels = c("Conventional","Organic"))+
  scale_x_continuous(breaks = c(2013:2019))+
  xlab("Year")+
  ylab("Average Field Size (ha)")+
  theme(legend.title = element_blank())

ggplot(mean_hectares, aes(x = year, y = mean_hectares))+
  geom_line(aes(colour = as.factor(cdfa_org)), size = 1)+
  geom_point(aes(colour = as.factor(cdfa_org)), size = 3)+
  theme_classic()+
  scale_colour_manual(values = c("darkgoldenrod2", "#2e4057"), labels = c("Conventional","Organic"))+
  scale_x_continuous(breaks = c(2013:2019))+
  xlab("Year")+
  ylim(10,40)+
  ylab("Average field size (ha)")+
  theme(legend.title = element_blank())
```

Soil Quality
```{r soilQuality}

mean_soil_noNA = all_years_df %>% 
  group_by(year,cdfa_org) %>% 
  summarise(mean_soilQ = round(mean(soil_noNA, na.rm=T),2))

mean_soil_NA = all_years_df %>% 
  group_by(year,cdfa_org) %>% 
  summarise(mean_soilQ = round(mean(soil_NA, na.rm=T),2))

# ggplot(mean_soilQ,aes(x = year,y = mean_soilQ,fill=as.factor(cdfa_org)))+
#   # geom_col(position = "dodge")+
#   # geom_label(aes(label = mean_hectares)) + 
#   theme_classic()+
#   scale_fill_manual(values = c("darkgoldenrod2", "#2e4057"), labels = c("Conventional","Organic"))+
#   scale_x_continuous(breaks = c(2013:2019))+
#   xlab("Year")+
#   ylab("Average Storie Index Soil Quality Rating \n [1 = highest]")+
#   theme(legend.title = element_blank())

# ggplot(all_years_df,aes(x = as.factor(cdfa_org),y = soil_quality))+
#   geom_violin()+
#   theme_classic()+
#   facet_wrap(~year)

ggplot(mean_soil_noNA, aes(x = year, y = mean_soilQ))+
  geom_line(aes(colour = as.factor(cdfa_org)), size = 1)+
  geom_point(aes(colour = as.factor(cdfa_org)))+
  theme_classic()+
  scale_colour_manual(values = c("darkgoldenrod2", "#2e4057"), labels = c("Conventional","Organic"))+
  scale_x_continuous(breaks = c(2013:2019))+
  xlab("Year")+
  ylim(1,3)+
  ylab("Average Storie Index Soil Quality Rating \n [1 = highest]")+
  theme(legend.title = element_blank())

ggplot(mean_soil_NA, aes(x = year, y = mean_soilQ))+
  geom_line(aes(colour = as.factor(cdfa_org)), size = 1)+
  geom_point(aes(colour = as.factor(cdfa_org)))+
  theme_classic()+
  scale_colour_manual(values = c("darkgoldenrod2", "#2e4057"), labels = c("Conventional","Organic"))+
  scale_x_continuous(breaks = c(2013:2019))+
  xlab("Year")+
  ylim(1,3)+
  ylab("Average Storie Index Soil Quality Rating \n [1 = highest]")+
  theme(legend.title = element_blank())
```
