==> Rmarkdown/8_Toxicity.Rmd <==
---
title: "8_Toxicity"
author: "Sofie McComb"
date: "4/27/2021"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

**This Rmarkdown document explores ways to generate summary ecotoxicological data for pesticides based on their CAS #s. In responses to reviewers we focus on results from the code chunk titles "aqEndpoints", but leave other code in this document for future reference and in case it is of interest to others.** 

Useful References:

Standartox Package:
https://www.master-ecotoxicology.de/ecotox-blog/standartox-standardizing-toxicity-data/
https://andschar.github.io/standartox/ 

EPA Toxicity Standards:
https://www.epa.gov/sites/production/files/2015-03/documents/chap-08-sep-2012.pdf
https://www3.epa.gov/pesticides/endanger/litstatus/effects/redleg-frog/rotenone/appendix-f.pdf 
https://www.epa.gov/pesticide-science-and-assessing-pesticide-risks/technical-overview-ecological-risk-assessment-0 
Bees: https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0094482 

Environmental Hazards: (Chp. 8)
"Based upon the results of eight basic acute toxicity studies performed on the technical grade of the active ingredient(s) in the formulation. These eight studies are: 
    (1) avian oral LD50 (with mallard or bobwhite quail), 
    (2) avian dietary LC50(mallards), 
    (3) avian dietary LC50 (bobwhite quail), 
    (4) freshwater fish LC50 (rainbow trout),
    (5) freshwater fish LC50 (bluegill sunfish), 
    (6) acute LC50 freshwater invertebrates (Daphnia magna or water flea), 
    (7) honeybee contact LD50, and 
    (8) mammalian acute oral LD50. 
Forspecific data requirements: 40 CFR Part 158."

1. Bird and Mammal Hazard Statement. The following statement has typically been
required when a pesticide intended for outdoor use contains an active ingredient which
has a mammalian acute oral LD50 ≤ 100 mg/kg, an avian acute oral LD50 ≤ 100 mg/kg,
or a subacute dietary LC50 ≤ 500 ppm:
“This pesticide is toxic to [birds] [mammals] or [birds and mammals]”.

2. Fish/Aquatic Invertebrate Statement. The following statement has typically been
required when a pesticide intended for outdoor use contains an active ingredient with a
fish acute LC50 or aquatic invertebrate (including estuarine species such as oyster and
mysid shrimp) EC50 ≤ 1 ppm:
“This pesticide is toxic to [fish] [fish and aquatic invertebrates]
[oysters/shrimp] or [fish, aquatic invertebrates, oysters and shrimp]”.

3. Incident Data Statement. If field studies or accident history, such as the FIFRA 6(a)(2)
reports, indicate that use of the pesticide may result in fatality to birds, fish or mammals,
the following statement has typically been required:
“This pesticide is extremely toxic to [birds], [mammals], [fish], or [birds and
mammals and fish]”.

Bees
Product contains any active ingredient with acute LD50 of 2 micrograms/bee or less 

This product is highly toxic to bees and other pollinating insects exposed to direct treatment or
residues on blooming crops or weeds. Do not apply this product or allow it to drift to blooming
crops or weeds if bees or other pollinating insects are visiting the treatment area

Product contains any active ingredient(s) with acute LD50 of greater than 2 micrograms/bee
but less than 11 micrograms/bee. This product is moderately toxic to bees and other pollinating insects
exposed to direct treatment or residues on blooming crops or weeds. Do not apply this product if
bees or other pollinating insects are visiting the treatment area.


Appendix F:
"In general, categories of acute toxicity ranging from “practically nontoxic” to “very highly toxic” have
been established for aquatic organisms based on LC50 values (Table F.1), terrestrial organisms based on
LD50 values (Table F.2), and avian species based on LD50 values (Table F.3). Subacute dietary toxicity
for avian species is based on the LC50 values (Table F.4). and non-target insects based on LD50 values
for honey bees (Table F.4) (EPA 2001)."

Fish: "The preferred test species are the bluegill sunfish, Lepomis macrochirus (a warmwater
fish), and the rainbow trout, Oncorhynchus mykiss (a coldwater fish)." 96-hour LC50

Daphnia: D. magna; 48-hour LC50 value

Birds: "The preferred test species is either Mallard duck (a waterfowl) or Bobwhite quail (an upland gamebird)."
Mallard duck (Anas platyrhynchos)
Ring-necked pheasant (Phasianus colchicus)

Bees: "The acute contact LD50, using the honey bee, Apis mellifera, is an acute
contact, single-dose laboratory study designed to estimate the quantity of toxicant required to cause 50%
mortality in a test population of bees. The TGAI is administered by one of two methods: whole body
exposure to technical pesticide in a nontoxic dust diluent or topical exposure to technical pesticide via
micro-applicator. The median LD50 is expressed in micrograms of active ingredient per bee (•g a.i./bee)."


https://www.ecfr.gov/cgi-bin/text-idx?SID=a2a0a7ef52614d357f2c9fd776dfb017&mc=true&node=pt40.26.158&rgn=div5#se40.26.158_1630

§158.630   Terrestrial and aquatic nontarget organisms data requirements table.
(3) In general, for all outdoor end-uses, including turf, the following studies are required: Two avian oral LD50, two avian dietary LC50, two avian reproduction studies, two freshwater fish LC50, one freshwater invertebrate EC50, one honeybee acute contact LD50, one freshwater fish early-life stage, one freshwater invertebrate life cycle, and three estuarine acute LC50/EC50 studies — fish, mollusk and invertebrate. All other outdoor residential uses, i.e., gardens and ornamental will not usually require the freshwater fish early-life stage, the freshwater invertebrate life-cycle, and the acute estuarine tests.
Fish: 13. Data are required on one freshwater fish species. If the test species is different from the two species used for the freshwater fish acute toxicity tests, a 96-hour LC50 on that species must also be provided.

```{r packages}

library(tidyverse)
# library(haven)
# library(pastecs)
options(scipen = 9999)
# install.packages('remotes') # if remotes:: is not installed
# remotes::install_github('andschar/standartox') # package not yet on CRAN
require(standartox)
require(data.table)
require(ggplot2)
library(PesticideLoadIndicator)
library(webchem)
rm(list=ls())

```


#### Aquatic Endpoints ####

Nowell data
```{r nowell}

nowell <- read_csv("./R_input/CSV/PesticideUse/NowellToPURSleuth.csv") %>% 
  filter(PurNowell=='matched (3)'|PurNowell=='master only (1)')
nowell_match <- filter(nowell,PurNowell=='matched (3)')

nonmissingcas <- c(nowell_match$casnum,fishdata$cas) %>% unique()

missingcas <- cas_data %>% 
  filter(!(casnum%in%nonmissingcas))

```


```{r aqEndpoints}

# Rainbow trout tests
trout_query = stx_query(endpoint = 'XX50',
              taxa = grep('Oncorhynchus mykiss', catal$taxa$variable, value = TRUE),
              exposure = 'aquatic',
              concentration_type = "active ingredient",
              effect="mortality",
              duration = 96)

troutdata = trout_query$filtered %>% 
  filter(cas%in%uniqcas) %>% 
  group_by(cas,concentration_unit) %>% 
  summarize(meanConc=mean(concentration,na.rm=T),
            medianConc=median(concentration,na.rm=T)) %>% 
  mutate(testType="trout_96hr_LC50")

# Blue gill tests
bluegill_query = stx_query(endpoint = 'XX50',
              taxa = grep('Lepomis macrochirus', catal$taxa$variable, value = TRUE), 
              exposure = 'aquatic',
              concentration_type = "active ingredient",
              effect="mortality",
              duration = 96)

bluegilldata = bluegill_query$filtered %>% 
  filter(cas%in%uniqcas) %>% 
  group_by(cas) %>% # don't need to group by concentration unit here. All either ppb or ug/l which are the same
  summarize(meanConc=mean(concentration,na.rm=T),
            medianConc=median(concentration,na.rm=T)) %>% 
  mutate(testType="bluegill_96hr_LC50")

# Aquatic invert tests
invert_query = stx_query(endpoint = 'XX50',
              taxa = grep('Daphnia magna', catal$taxa$variable, value = TRUE), 
              exposure = 'aquatic',
              concentration_type = "active ingredient",
              effect="mortality",
              duration = 48)

invertdata = invert_query$filtered %>% 
  filter(cas%in%uniqcas) %>% 
  group_by(cas,concentration_unit) %>% # don't need to group by concentration unit here. All either ppb or ug/l which are the same
  summarize(meanConc=mean(concentration,na.rm=T),
            medianConc=median(concentration,na.rm=T)) %>% 
  mutate(testType="aqinvert_48hr_lc50")


aqsppdata <- rbind(troutdata,bluegilldata,invertdata)
cas_data_wTox <- merge(cas_data,aqsppdata,by.x="casnum",by.y="cas",all.x=T)

# fish invert tests
fish_query = stx_query(endpoint = 'XX50',
              # taxa = grep('Daphnia magna', catal$taxa$variable, value = TRUE), 
              exposure = 'aquatic',
              concentration_type = "active ingredient",
              ecotox_grp = "fish",
              effect="mortality",
              duration = 96)

#### Any data from 96 hr LC 50 aquatic tests ####

fishdata = fish_query$filtered %>% 
  filter(cas%in%uniqcas) %>% 
  group_by(cas) %>% # Concentration unit as grouping feature here
  summarize(#meanConc=mean(concentration,na.rm=T),
            medianConc=median(concentration,na.rm=T)) %>% 
  mutate(testType="fish_96hr_lc50")
colnames(fishdata)[1]="casnum"

fishNowell <- left_join(nowell,fishdata,by="casnum") %>% 
  mutate(NowellWebchemDiff = (MTCugL-medianConc))

write_csv(fishNowell, "R_input/CSV/PesticideUse/FishTox.csv")
```


########


#### Trying out webchem and standartox package ####

```{r CASdata}
### Read in cas data from Ashley
cas_data<-read_csv("./R_input/CSV/PesticideUse/KernChemsUsed1319_Cas.csv")
casdf<-cas_data %>% 
  filter(!is.na(casnum))
uniqcas<-unique(casdf$casnum)

#Read in standartox catalog
require(standartox)
catal = stx_catalog()
# names(catal)

```

# Find CAS #s with valid ecotox data using webchem, then see if possible to find errors in non-matching cas #s
```{r getEcotox}
# # Read in CAS CSV data from Ashley
# cas_data<-read_csv("./R_input/CSV/PesticideUse/KernChemsUsed1319_Cas.csv") 
# 
# # Create a new column of NAs, where boolean for valid cas will be stored
# cas_data <- cas_data %>% mutate(validcas=NA)
# etoxs <- NULL
# for(i in 1:nrow(cas_data)){
#   
#   cas <- cas_data$casnum[i]
#   cas_data$validcas[i] <- is.cas(cas)
#   etox <- get_etoxid(cas,from = c("cas"))
#   etoxs <- rbind(etoxs,etox)
#   rm(cas,etox)
#   }
# 
# # Filter CAS data to keep those without a valid ecotox ID match
# casdata_nomatch <- cas_data %>% filter(is.na(validcas)|validcas==FALSE)
# 
# # See if we can find matches with names for those
# casdata_namematch <- NULL
# for(i in 1:nrow(casdata_nomatch )){
#   
#   name <- tolower(casdata_nomatch$chemname[i])
#   cas <- tolower(casdata_nomatch$casnum[i])
#   etox <- get_etoxid(query = name,from = c("name"))
#   casdata_namematch <- rbind(casdata_namematch,cbind(etox,cas))
# }
# 
# colnames(etoxs)[1]<-"casnum"
# 
# chemetox <- inner_join(chems,etoxs,by="casnum") %>% unique()
# df <- as.data.frame(table(chemetox$validcas,chemetox$PURcas))
```


####### Sofie's tox code #######

(1) freshwater fish LC50 (rainbow trout),

```{r RainbowTroutLC50}

#Oncorhynchus mykiss (a coldwater fish)--Rainbow Trout--96-hour LC50

# Query
onmy_df<-data.frame()
onmy_agg<-data.frame()
for(i in 1:length(uniqcas)){
  id<-uniqcas[i]
  id_query = stx_query(casnr = id,
              endpoint = 'XX50',
              taxa = grep('Oncorhynchus mykiss', catal$taxa$variable, value = TRUE), # fish genus
              exposure = 'aquatic',
              concentration_type = "active ingredient",
              effect="mortality",
              duration = 96
              )
  onmy_df<-rbind(onmy_df, id_query$filtered_all, fill=TRUE)
  onmy_agg<-rbind(onmy_agg, id_query$aggregated, fill=TRUE)
}

write_csv(onmy_df, "../R_input/CSV/Toxicity/metrics/onmy_df.csv")
write_csv(onmy_agg, "../R_input/CSV/Toxicity/metrics/onmy_agg.csv")

```


(2) freshwater fish LC50 (bluegill sunfish), 

```{r BluegillLC50}

#bluegill sunfish, Lepomis macrochirus (a warmwaterfish)

# Query
lema_df<-data.frame()
lema_df_agg<-data.frame()
for(i in 1:length(uniqcas)){
  id<-uniqcas[i]
  id_query = stx_query(casnr = id,
              endpoint = 'XX50',
              taxa = grep('Lepomis macrochirus', catal$taxa$variable, value = TRUE),
              exposure = 'aquatic',
              concentration_type = "active ingredient",
              effect="mortality",
              duration = c(0,96)
              )
  lema_df<-rbind(lema_df, id_query$filtered_all, fill=TRUE)
  lema_df_agg<-rbind(lema_df_agg, id_query$aggregated, fill=TRUE)
}

write_csv(lema_df, "../R_input/CSV/Toxicity/metrics/lema_df.csv")
write_csv(lema_df_agg, "../R_input/CSV/Toxicity/metrics/lema_df_agg.csv")

```


(3) acute LC50 freshwater invertebrates (Daphnia magna or water flea), 
https://www.epa.gov/pesticide-science-and-assessing-pesticide-risks/technical-overview-ecological-risk-assessment-0
Invert section

```{r invertsLC50}

#acute LC50 freshwater invertebrates (Daphnia magna or water flea)

# Query
dama_df<-data.frame()
dama_df_agg<-data.frame()
for(i in 1:length(uniqcas)){
  id<-uniqcas[i]
  id_query = stx_query(casnr = id,
              endpoint = 'XX50',
              taxa = grep('Daphnia magna', catal$taxa$variable, value = TRUE),
              exposure = 'aquatic',
              concentration_type = "active ingredient",
              effect="mortality",
              duration=c(0,48)
              )
  dama_df<-rbind(dama_df, id_query$filtered_all, fill=TRUE)
  dama_df_agg<-rbind(dama_df_agg, id_query$aggregated, fill=TRUE)
}

write_csv(dama_df, "../R_input/CSV/Toxicity/metrics/dama_df.csv")
write_csv(dama_df_agg, "../R_input/CSV/Toxicity/metrics/dama_df_agg.csv")

```


(4) avian oral LD50 (with mallard or bobwhite quail)

No oral mallard or bobwhite studies found


(5) avian dietary LC50(mallards), 

```{r}

# #Mallards (Anas platyrhynchos) LD50
# 
# # Query
# anpl_df<-data.frame()
# anpl_df_agg<-data.frame()
# for(i in 1:length(uniqcas)){
#   id<-uniqcas[i]
#   id_query = stx_query(casnr = id,
#               endpoint = 'XX50',
#               taxa = grep('Anas platyrhynchos', catal$taxa$variable, value = TRUE),
#               exposure = 'diet',
#               concentration_type = "active ingredient",
#               effect="mortality",
#               duration = c(0,192) #8 day lab study
#               )
#   anpl_df<-rbind(anpl_df, id_query$filtered_all, fill=TRUE)
#   anpl_df_agg<-rbind(anpl_df_agg, id_query$aggregated, fill=TRUE)
# }
# 
# write_csv(anpl_df, "../R_input/CSV/Toxicity/metrics/anpl_df.csv")
# write_csv(anpl_df_agg, "../R_input/CSV/Toxicity/metrics/anpl_df_agg.csv")

```


(6) avian dietary LC50 (bobwhite quail)

```{r}
# 
# #Colinus virginianus Bobwhite Quails LD50
# 
# # Query
# covi_df<-data.frame()
# covi_df_agg<-data.frame()
# for(i in 1:length(uniqcas)){
#   id<-uniqcas[i]
#   id_query = stx_query(casnr = id,
#               endpoint = 'XX50',
#               taxa = grep('Colinus virginianus', catal$taxa$variable, value = TRUE),
#               exposure = 'diet',
#               concentration_type = "active ingredient",
#               effect="mortality",
#               duration = c(0,192)#8 day lab study
#               )
#   covi_df<-rbind(covi_df, id_query$filtered_all, fill=TRUE)
#   covi_df_agg<-rbind(covi_df_agg, id_query$aggregated, fill=TRUE)
# }
# 
# write_csv(covi_df, "../R_input/CSV/Toxicity/metrics/covi_df.csv")
# write_csv(covi_df_agg, "../R_input/CSV/Toxicity/metrics/covi_df_agg.csv")

```


(7) honeybee contact LD50, and 
https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0094482 

```{r honeybeeLC50}
# 
# #Apis mellifera honeybee LD50
# 
# # Query
# apme_df<-data.frame()
# apme_df_agg<-data.frame()
# for(i in 1:length(uniqcas)){
#   id<-uniqcas[i]
#   id_query = stx_query(casnr = id,
#               endpoint = 'XX50',
#               taxa = grep('Apis mellifera', catal$taxa$variable, value = TRUE),
#               exposure = 'topical', #contact, acute
#               concentration_type = "active ingredient",
#               effect="mortality",
#               duration = c(0,48) #Within 2-day period
#               )
#   apme_df<-rbind(apme_df, id_query$filtered_all, fill=TRUE)
#   apme_df_agg<-rbind(apme_df_agg, id_query$aggregated, fill=TRUE)
# }
# 
# write_csv(apme_df, "../R_input/CSV/Toxicity/metrics/apme_df.csv")
# write_csv(apme_df_agg, "../R_input/CSV/Toxicity/metrics/apme_df_agg.csv")

```


(8) mammalian acute oral LD50.
Not examining mammals in our study

```{r AllTox}

#Combine aggregated values in CAS dataset
casdf  <- cas_data %>% mutate(cas=casnum)
# AvianDietMallardLC50<-anpl_df_agg%>% 
#   mutate(AvianDietMallardLC50=gmn) %>% 
#   select(cas, AvianDietMallardLC50) %>% 
#   na.omit()
# AvianDietQuailLC50<-covi_df_agg%>% 
#   mutate(AvianDietQuailLC50=gmn) %>% 
#   select(cas, AvianDietQuailLC50) %>% 
#   na.omit()
FishTroutLC50<-onmy_agg%>% 
  mutate(FishTroutLC50=gmn) %>% 
  select(cas, FishTroutLC50) %>% 
  na.omit()
FishSunfishLC50<-lema_df_agg%>% 
  mutate(FishSunfishLC50=gmn) %>% 
  select(cas, FishSunfishLC50) %>% 
  na.omit()
InvertDaphniaLC50<-dama_df_agg%>% 
  mutate(InvertDaphniaLC50=gmn) %>% 
  select(cas, InvertDaphniaLC50) %>% 
  na.omit()
# BeeLD50<-apme_df_agg%>% 
#   mutate(BeeLD50=gmn) %>% 
#   select(cas, BeeLD50) %>% 
#   na.omit()

#Join all casdf with matching Tox data
cas_tox_df <- list(casdf,
                   FishTroutLC50,
                   FishSunfishLC50,
                   InvertDaphniaLC50) %>% 
  reduce(left_join, by = "cas")

write_csv(cas_tox_df, "../R_input/CSV/Toxicity/cas_tox_df.csv")


```
