==> Rmarkdown/10.2_OLS_Claire.Rmd <==
---
title: "Claire: OLS"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```


#### 1. Load packages, set working directory, and prevent scientific notation
```{r get_packages, message=FALSE, warning=FALSE}
library(tidyverse)
library(rgdal)
library(sf)
library(MatchIt) ## Package with Propensity Score Matching
library(MASS)
options(scipen = 9999) ## Avoid scientific notation
```

Read in data
```{r data}

data = read_csv("../R_output/CSV/compiled_organics1317.csv")

soil = data %>% 
  group_by(all_org) %>% 
  summarise(mean_soil = mean(soil_quality),
            total_fiels = n())

soil_fltr = data %>% 
  filter(soil_quality>0) %>% 
  group_by(all_org) %>% 
  summarise(mean_soil = mean(soil_quality),
            total_fiels = n())

data$full_organic = ifelse(data$KernPermittee_org == 1,1,data$all_org)
org_col = 'full_organic'

farm_stats_df = data %>% 
  # filter(year == 2017) %>% 
  group_by(permit,permittee) %>% 
  summarise(total_org = sum(eval(as.name(org_col))),
            total_conv = n()-sum(eval(as.name(org_col))),
            average_fldsize_org = round(mean(hectares[eval(as.name(org_col)) == 1]),2),
            average_fldsize_conv = round(mean(hectares[eval(as.name(org_col)) == 0]),2),
            average_soil_org = round(mean(soil_quality[eval(as.name(org_col)) == 1]),2),
            average_soil_conv = round(mean(soil_quality[eval(as.name(org_col)) == 0]),2)) %>% 
  filter(total_org>0&total_conv>0)

## Combing growers across years
crop_stats_df = data %>% 
  group_by(comm_edit) %>% 
  summarise(total_fields = n(),
    total_org = sum(eval(as.name(org_col))),
    total_conv = n()-sum(eval(as.name(org_col)))) %>% 
  filter(total_fields > 99 & total_org > 49)
  
agro_stats_df = data %>% 
  group_by(agroclass) %>% 
  summarise(total_fields = n(),
    total_org = sum(eval(as.name(org_col))),
    total_conv = n()-sum(eval(as.name(org_col)))) %>% 
  filter(total_fields>99 & total_org > 49)

```

```{r}
data$permit = as.factor(data$permit)
lm_test = lm(KgAIperHE ~ soil_quality + all_org + permit, data = data)

dataNA = data
dataNA$soil_quality[dataNA$soil_quality<1] = NA
lm_testNA = lm(KgAIperHE ~ soil_quality + all_org + permit, data = dataNA)

data_fltr = data %>% 
  filter(!is.na(soil_quality))
lm_test_fltr = lm(KgAIperHE ~ soil_quality + all_org + permit, data = data_fltr)

table(data$year)-table(data_fltr$year)

summary(lm_test)
summary(lm_testNA)
summary(lm_test_fltr)
```
